+++ src/App.js
@@ -4,8 +4,7 @@ import { Auth0Provider } from '@auth0/auth0-react';
 import { auth0Config } from './utils/auth0Config';
 import './App.css';
 import Navigation from './components/Navigation';
-import InvoicesDash from './components/Invoices/InvoicesDash';
-import InvoicesGen from './components/Invoices/InvoicesGen';
+import { InvoicesDash, InvoicesGen } from './pages/Invoices/index';
 import EstimatesGen from './components/Estimates/EstimatesGen';
 import EstimatesDash from './components/Estimates/EstimatesDash';
 import { ClientsDash, ClientsGen } from './pages/Clients/index';
