+++ src/pages/Invoices/index.js
@@ -1,3 +1,2 @@
 export { default as InvoicesDash } from './InvoicesDash';
 export { default as InvoicesGen } from './InvoicesGen';
-export { default as InvoicesTemplate } from './InvoicesTemplate';
